+++ client/catapult/tools/linker/main.cpp
@@ -57,10 +57,10 @@ namespace catapult { namespace tools { namespace linker {
 						OptionsValue<std::string>(),
 						"linked public key (32 bytes for vrf, 48 bytes for voting)");
 				optionsBuilder("startEpoch,b",
-						OptionsValue<uint64_t>()->default_value(1),
+						OptionsValue<uint32_t>()->default_value(1),
 						"voting key start epoch");
 				optionsBuilder("endEpoch,e",
-						OptionsValue<uint64_t>()->default_value(26280),
+						OptionsValue<uint32_t>()->default_value(26280),
 						"voting key end epoch");
 				optionsBuilder("output",
 						OptionsValue<std::string>(),
@@ -75,8 +75,8 @@ namespace catapult { namespace tools { namespace linker {
 				auto networkIdentifier = config.BlockChain.Network.Identifier;
 				auto signer = crypto::KeyPair::FromString(options["secret"].as<std::string>());
 				auto linkedPublicKey = options["linkedPublicKey"].as<std::string>();
-				auto startEpoch = FinalizationEpoch(options["startEpoch"].as<uint64_t>());
-				auto endEpoch = FinalizationEpoch(options["endEpoch"].as<uint64_t>());
+				auto startEpoch = FinalizationEpoch(options["startEpoch"].as<uint32_t>());
+				auto endEpoch = FinalizationEpoch(options["endEpoch"].as<uint32_t>());
 				auto pTransaction = options["type"].as<std::string>() == "voting"
 						? createVotingKeyLinkTransaction(networkIdentifier, signer.publicKey(), linkedPublicKey, startEpoch, endEpoch)
 						: createVrfKeyLinkTransaction(networkIdentifier, signer.publicKey(), linkedPublicKey);
